#2 added editor config and cleaned up space for editor file
